Updated the New function to accept a s3 key prefix option ... (via functional options)
Fixed the log method to correctly expand the passed in variadic interface

The `New` method now uses functional options to pass in additional data. There is now an exported `KeyPrefix` functional option that accepts a key prefix. There is also a non-exported `s3session` functional option that is used when testing to override the s3 session creation.